@@ -19,19 +19,14 @@
   "scripts": {
     "test": "npm run build-prod && cross-env NODE_ENV=test mocha --opts mocha.opts",
     "build-prod": "rimraf dist && cross-env NODE_ENV=production npx tsc",
-    "lint": "eslint src test",
+    "lint": "npx eslint -c .eslintrc.js src/**/*.ts --fix",
     "generate-doc": "npx esdoc"
   },
   "dependencies": {
-    "@types/i18n": "^0.8.3",
-    "@types/kcors": "^2.2.3",
-    "@types/koa": "^2.0.48",
-    "@types/koa-bodyparser": "^4.2.1",
-    "@types/koa-router": "^7.0.39",
-    "@types/node": "^10.12.24",
     "apidoc": "^0.17.7",
     "babel-runtime": "^6.26.0",
     "chalk": "^2.4.2",
+    "esdoc-typescript-plugin": "^1.0.1",
     "fs-extra": "^8.0.1",
     "i18n-2": "^0.7.3",
     "jsonwebtoken": "^8.5.1",
@@ -50,28 +45,37 @@
     "node-gettext": "^2.0.0"
   },
   "devDependencies": {
+    "@types/i18n": "^0.8.3",
+    "@types/kcors": "^2.2.3",
+    "@types/koa": "^2.0.48",
+    "@types/koa-bodyparser": "^4.2.1",
+    "@types/koa-router": "^7.0.39",
+    "@types/node": "^10.14.17",
+    "@typescript-eslint/eslint-plugin": "^2.3.2",
+    "@typescript-eslint/eslint-plugin-tslint": "^2.3.2",
+    "@typescript-eslint/parser": "^2.3.2",
     "awesome-typescript-loader": "^5.2.1",
     "babel-cli": "^6.26.0",
-    "babel-eslint": "^10.0.1",
+    "babel-eslint": "^10.0.3",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.5",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-env": "^1.7.0",
     "babel-preset-stage-3": "^6.24.1",
-    "cross-env": "^5.2.0",
+    "cross-env": "^5.2.1",
     "esdoc": "^1.1.0",
     "esdoc-ecmascript-proposal-plugin": "^1.0.0",
     "esdoc-standard-plugin": "^1.0.0",
     "eslint": "^5.16.0",
     "expect": "^24.8.0",
     "mocha": "^6.1.4",
-    "nodemon": "^1.19.1",
+    "nodemon": "^1.19.2",
     "prettier-eslint": "^8.8.2",
     "rimraf": "^2.6.3",
-    "supertest": "^4.0.2"
     "source-map-loader": "^0.2.4",
-    "supertest": "^3.3.0",
-    "typescript": "^3.3.1"
+    "supertest": "^4.0.2",
+    "tslint": "^5.20.0",
+    "typescript": "^3.6.2"
   },
   "keywords": [
     "koa-smart",
