@@ -1,10 +1,9 @@
 import moment from 'moment';
 import jsonwebtoken from 'jsonwebtoken';
-
 import * as utils from './utils';
 
-module.exports = {
-  jsonwebtoken,
+export default {
   moment,
-  utils,
+  jsonwebtoken,
+  ...utils,
 };
