@@ -1,13 +1,25 @@
 const __ = string => string;
 
-/**
- * @typedef OptionErrors
- * @property {boolean} [logAll = false] if set to true, all logs will be displayed regardless of their individual settings
- * @property {boolean} [logErrorUnknown = false] whether unknown errors should be displayed
- * @property {boolean} [logErrorSequelize = false] whether errors pertaining to the models should be logged
- * @property {boolean} [logErrorApp = false] whether errors coming from thrown {@link ErrorApp} should be logged
- */
-let options = {
+interface OptionsErrors {
+  /**
+   * if set to true, all logs will be displayed regardless of their individual settings
+   */
+  logAll: boolean;
+  /**
+   * whether unknown errors should be displayed
+   */
+  logErrorUnknown: boolean;
+  /**
+   * whether errors pertaining to the models should be logged
+   */
+  logErrorSequelize: boolean;
+  /**
+   * whether errors coming from thrown {@link ErrorApp} should be logged
+   */
+  logErrorApp: boolean;
+}
+
+let options: OptionsErrors = {
   logAll: false,
   logErrorUnknown: false,
   logErrorSequelize: false,
@@ -36,7 +48,10 @@ async function handleError(ctx, next) {
   try {
     await next();
   } catch (err) {
-    const arraySequelize = ['SequelizeValidationError', 'SequelizeUniqueConstraintError'];
+    const arraySequelize = [
+      'SequelizeValidationError',
+      'SequelizeUniqueConstraintError',
+    ];
 
     if (err.constructor.name === 'ErrorApp') {
       // expected error
@@ -45,25 +60,39 @@ async function handleError(ctx, next) {
       if (err.messages) {
         ctx.body.messages = err.messages;
       } else {
-        ctx.body.message = getMessageTranslate(ctx, err.message, err.toTranslate);
+        ctx.body.message = getMessageTranslate(
+          ctx,
+          err.message,
+          err.toTranslate
+        );
       }
       displayLog(err, 'logErrorApp');
     } else if (arraySequelize.includes(err.name)) {
       // sequilize expected error by validattor or other
       ctx.status = 400;
-      ctx.body = { message: getMessageTranslate(ctx, err.message.split(':').pop(), true) };
+      ctx.body = {
+        message: getMessageTranslate(ctx, err.message.split(':').pop(), true),
+      };
       displayLog(err, 'logErrorSequelize');
     } else {
       // unexpected error
       ctx.status = 500;
-      ctx.body = { message: getMessageTranslate(ctx, __('Please contact the support'), true) };
+      ctx.body = {
+        message: getMessageTranslate(
+          ctx,
+          __('Please contact the support'),
+          true
+        ),
+      };
       displayLog(err, 'logErrorUnknown');
     }
   }
 }
 
 /**
- * middleware in charge of handling errors thrown on purpose, either through manually throwing {@link ErrorApp}, either through calling {@link Route.throw}.
+ * middleware in charge of handling errors thrown on purpose,
+ * either through manually throwing {@link ErrorApp},
+ * either through calling {@link Route.throw}.
  *
  * It will also make sure errors pertaining to models as well as unexpected error are given a clearer message.
  * @param {OptionErrors} [opt = {}] option object to set which events should be logged
