@@ -1,23 +1,57 @@
-/**
- * @typedef {Object} ParamsMethodDecorator
- * @property {Types} bodyType Description of the body receiving
- * @property {Types} queryType Description of the "request.query" receiving
- * @property {string} path the path at which the route will be available.<br/>
- * By default it will take the name of the function and replace uppercase by "-".<br/>
- * ex: a function name addUser will be mount with /add-user
- * @property {boolean} disabled if set to true, the route will be ignored
- * @property {function[]} middlewares an array of Koa Middlewares, which will be mounted for the given route
- * @property {Object} rateLimit a rateLimit object, which lets the user describe the max rate at which a user can access the route
- * @property {function[]} accesses an array of async function, which will be call with ctx, if one of them return true, the current client will access the route. This will overwrite the accesses pass to {ParamsClassDecorator}
- */
+export interface ParamsMethodDecorator {
+  /**
+   * Description of the body receiving
+   */
+  bodyType: any;
+  /**
+   * Description of the "request.query" receiving
+   */
+  queryType: any;
+  /**
+   * the path at which the route will be available.
+   * By default it will take the name of the function and replace uppercase by "-".
+   * Ex: a function name addUser will be mount with /add-user
+   */
+  path: string;
+  /**
+   * if set to true, the route will be ignored
+   */
+  disabled: boolean;
+  /**
+   * an array of Koa Middlewares, which will be mounted for the given route
+   */
+  middlewares: Function[];
+  /**
+   * a rateLimit object, which lets the user describe the max rate at which a user can access the route
+   */
+  rateLimit: Object;
+  /**
+   * an array of async function, which will be call with ctx,
+   * if one of them return true, the current client will access the route.
+   * This will overwrite the accesses pass to {ParamsClassDecorator}
+   */
+  accesses: Function[];
+}
 
-/**
- * @typedef {Object} ParamsClassDecorator
- * @property {string} routeBase a prefix which will be preppended all to the route's path
- * @property {boolean} disabled if set to true, all route in the class will be ignored
- * @property {function[]} middlewares an array of Koa Middlewares, which will be mounted for the given route
- * @property {function[]} accesses an array of async function, which will be call (for all routes in the class) with ctx, if one of them return true, the current client will access the route
- */
+export interface ParamsClassDecorator {
+  /**
+   * a prefix which will be preppended all to the route's path
+   */
+  routeBase: string;
+  /**
+   * if set to true, all route in the class will be ignored
+   */
+  disabled: boolean;
+  /**
+   * an array of Koa Middlewares, which will be mounted for the given route
+   */
+  middlewares: Function[];
+  /**
+   * an array of async function, which will be call (for all routes in the class) with ctx,
+   * if one of them return true, the current client will access the route
+   */
+  accesses: Function[];
+}
 
 /**
  *@ignore
@@ -25,9 +59,9 @@
 export default class RouteDecorators {
   // replace A-Z to lowercase and add - ex myFunc => my-func
   static _getRouteFromMethode(str) {
-    return str.replace(/([A-Z])/g, (str, letter, index) => {
-      return index === 0 ? `${letter.toLowerCase()}` : `-${letter.toLowerCase()}`;
-    });
+    return str.replace(/([A-Z])/g, (fullString, letter, index) =>
+      index === 0 ? `${letter.toLowerCase()}` : `-${letter.toLowerCase()}`
+    );
   }
 
   static _initData(target) {
@@ -72,7 +106,9 @@ export default class RouteDecorators {
       Object.defineProperty(target, functionName, descriptor);
 
       const routeName =
-        options.path === undefined ? RouteDecorators._getRouteFromMethode(functionName) : options.path;
+        options.path === undefined
+          ? RouteDecorators._getRouteFromMethode(functionName)
+          : options.path;
 
       target.routeBase = options.routeBase;
       target.routes[type].push({
