@@ -1,25 +1,28 @@
 import uniq from 'lodash.uniq';
 
 import { TypeAny } from './TypeAny';
-import { utils } from '../utils';
+import utils from '../utils';
 
 export class TypeBoolean extends TypeAny {
   _truthyValues = ['true']; // Specifies additional values to be considered as 'truthy'
   _falsyValues = ['false']; // Specifies additional values to be considered as 'falsy'
   _insensitive = true;
 
-  constructor(params = {}) {
+  constructor(params = { i18n: {} }) {
     super({ ...params, type: 'boolean' });
   }
 
-  _getErrorInvalidValue = ({ _i18n }) => {
+  _getErrorInvalidValue({ _i18n }) {
     return _i18n.__('Should be a boolean');
-  };
+  }
 
-  _getDescription = (prefix = 'It should be ') => {
+  _getDescription(prefix = 'It should be ') {
     const valideValue = [...this._truthyValues, ...this._falsyValues];
-    return `${prefix}a boolean or one of: (${utils.joinWithCote(valideValue, ', ')}).`;
-  };
+    return `${prefix}a boolean or one of: (${utils.joinWithCote(
+      valideValue,
+      ', '
+    )}).`;
+  }
 
   _insensitiveArray(array) {
     return array.map(value => {
@@ -31,7 +34,10 @@ export class TypeBoolean extends TypeAny {
   }
 
   truthy(vals = []) {
-    this._truthyValues = uniq([...this._truthyValues, ...utils.castArray(vals)]);
+    this._truthyValues = uniq([
+      ...this._truthyValues,
+      ...utils.castArray(vals),
+    ]);
     return this;
   }
 
@@ -48,13 +54,17 @@ export class TypeBoolean extends TypeAny {
   _testType() {
     if (!['boolean', 'string', 'number'].includes(typeof this._value)) {
       this._setError(this._TypeError.INVALID_TYPE);
+      return false;
     }
+    return true;
   }
 
   _test() {
     if (typeof this._value !== 'boolean') {
       this._setError(this._TypeError.INVALID_VALUE);
+      return false;
     }
+    return true;
   }
 
   _transform() {
