@@ -1,22 +1,25 @@
 import { TypeAny } from './TypeAny';
-import { utils } from '../utils';
+import utils from '../utils';
 
 export class TypeEnum extends TypeAny {
-  _oneOf = [];
+  _oneOf: Array<any> = [];
   _insensitive = true;
   _number = true;
 
-  constructor(params = {}) {
+  constructor(params = { i18n: {} }) {
     super({ ...params, type: 'enum' });
   }
 
-  _getErrorInvalidValue = ({ _i18n }) => {
-    return _i18n.__('Should be one of %s', utils.joinWithCote(this._oneOf, ', '));
-  };
+  _getErrorInvalidValue({ _i18n }) {
+    return _i18n.__(
+      'Should be one of %s',
+      utils.joinWithCote(this._oneOf, ', ')
+    );
+  }
 
-  _getDescription = (prefix = 'It should be ') => {
+  _getDescription(prefix = 'It should be ') {
     return `${prefix}one of: (${utils.joinWithCote(this._oneOf, ', ')}).`;
-  };
+  }
 
   _insensitiveArray(array) {
     return array.map(value => {
@@ -45,19 +48,26 @@ export class TypeEnum extends TypeAny {
   _testType() {
     if (!['string', 'number'].includes(typeof this._value)) {
       this._setError(this._TypeError.INVALID_TYPE);
+      return false;
     }
+    return true;
   }
 
   _test() {
     if (!this._oneOf.includes(this._value)) {
       this._setError(this._TypeError.INVALID_VALUE);
+      return false;
     }
+    return true;
   }
 
   _transform() {
     if (this._insensitive && typeof this._value === 'string') {
       for (const e of this._oneOf) {
-        if (typeof e === 'string' && e.toLocaleLowerCase() === this._value.toLocaleLowerCase()) {
+        if (
+          typeof e === 'string' &&
+          e.toLocaleLowerCase() === this._value.toLocaleLowerCase()
+        ) {
           this._value = e;
         }
       }
