@@ -10,31 +10,44 @@ export class TypeNumber extends TypeAny {
 
   _tPrecision; // Specifies the maximum number of decimal places where:
   _tPrecisionType; // Specifies the type of precision : floor, ceil, trunc, round
+  _errorKey;
 
-  constructor(params = {}) {
+  constructor(params = { i18n: {} }) {
     super({ ...params, type: 'number' });
   }
 
-  _getErrorInvalidValue = ({ _i18n }, key) => {
+  _getErrorInvalidValue({ _i18n }, key) {
     key = this._errorKey || key;
     this._errorKey = key;
 
-    if (key === 'between') return _i18n.__('Is not between %d and %d', this._min, this._max);
-    if (key === 'min') return _i18n.__('Is smaller than %d', this._min);
-    if (key === 'max') return _i18n.__('Is greater than %d', this._max);
-    if (key === 'multiple') return _i18n.__('Is not a multiple of %d', this._multiple);
-    if (key === 'positive') return _i18n.__('Is negative');
-    if (key === 'negative') return _i18n.__('Is positive');
+    if (key === 'between') {
+      return _i18n.__('Is not between %d and %d', this._min, this._max);
+    }
+    if (key === 'min') {
+      return _i18n.__('Is smaller than %d', this._min);
+    }
+    if (key === 'max') {
+      return _i18n.__('Is greater than %d', this._max);
+    }
+    if (key === 'multiple') {
+      return _i18n.__('Is not a multiple of %d', this._multiple);
+    }
+    if (key === 'positive') {
+      return _i18n.__('Is negative');
+    }
+    if (key === 'negative') {
+      return _i18n.__('Is positive');
+    }
     return null;
-  };
+  }
 
-  _getDescription = (prefix = 'It should be ') => {
+  _getDescription(prefix = 'It should be ') {
     let pN = ' ';
     pN = this._positive ? ' positive ' : pN;
     pN = !this._positive && this._negative ? ' negative ' : pN;
-    let msgError = `${prefix}a${pN}number`;
+    const msgError = `${prefix}a${pN}number`;
 
-    const paramsDesc = [];
+    const paramsDesc: Array<string> = [];
     if (this._min != null && this._max != null) {
       paramsDesc.push(`is between ${this._min} and ${this._max}`);
     } else if (this._min != null) {
@@ -45,8 +58,11 @@ export class TypeNumber extends TypeAny {
     if (this._multiple != null) {
       paramsDesc.push(`is a multiple of ${this._multiple}`);
     }
-    return `${msgError}${this._generateParamDescription(paramsDesc, ' which')}.`;
-  };
+    return `${msgError}${this._generateParamDescription(
+      paramsDesc,
+      ' which'
+    )}.`;
+  }
 
   min(nb = 0) {
     this._min = nb;
@@ -99,31 +115,53 @@ export class TypeNumber extends TypeAny {
   _testType() {
     if (!this._isNumber()) {
       this._setError(this._TypeError.INVALID_TYPE);
+      return false;
     }
+    return true;
   }
 
   _test() {
     const t = this._TypeError.INVALID_VALUE;
     const tMin = this._min != null && this._value < this._min;
     const tMax = this._max != null && this._value > this._max;
-    if (tMin && tMax) return this._setError(t, 'between');
-    if (tMin) return this._setError(t, 'min');
-    if (tMax) return this._setError(t, 'max');
-    if (this._multiple != null && this._value % this._multiple !== 0) return this._setError(t, 'multiple');
-    if (this._positive && this._value < 0) return this._setError(t, 'positive');
-    if (this._negative && this._value >= 0) return this._setError(t, 'negative');
+    if (tMin && tMax) {
+      return this._setError(t, 'between');
+    }
+    if (tMin) {
+      return this._setError(t, 'min');
+    }
+    if (tMax) {
+      return this._setError(t, 'max');
+    }
+    if (this._multiple != null && this._value % this._multiple !== 0) {
+      return this._setError(t, 'multiple');
+    }
+    if (this._positive && this._value < 0) {
+      return this._setError(t, 'positive');
+    }
+    if (this._negative && this._value >= 0) {
+      return this._setError(t, 'negative');
+    }
+    return true;
   }
 
-  _precisionTo = (nb, nbDigit, type) => Math[type](nb * 10 ** nbDigit) / 10 ** nbDigit;
+  _precisionTo(nb, nbDigit, type) {
+    return Math[type](nb * 10 ** nbDigit) / 10 ** nbDigit;
+  }
 
   _transform() {
     if (this._integer) {
-      this._value = parseInt(this._value);
+      this._value = parseInt(this._value, 10);
     } else {
       this._value = parseFloat(this._value);
     }
 
-    if (this._tPrecision >= 0)
-      this._value = this._precisionTo(this._value, this._tPrecision, this._tPrecisionType);
+    if (this._tPrecision >= 0) {
+      this._value = this._precisionTo(
+        this._value,
+        this._tPrecision,
+        this._tPrecisionType
+      );
+    }
   }
 }
